move classes to their own files as they were starting to get quite large
