+++ research/blueprint_algo/player.py
@@ -1,0 +1,17 @@
+from pluribus.game.player import Player
+from pluribus.game.pot import Pot
+
+
+class ShortDeckPokerPlayer(Player):
+    """Inherits from Player which will interface easily with the PokerEngine.
+
+    This class should manage the state of the players personal pot, and the
+    private cards that are dealt to the player. Also manages whether this
+    player has folded or not.
+    """
+
+    def __init__(self, player_i: int, initial_chips: int, pot: Pot):
+        """Instanciate a player."""
+        super().__init__(
+            name=f"player_{player_i}", initial_chips=initial_chips, pot=pot,
+        )
